ref #10 : added example program & moved code inside a specific folder
